--- interactions/View.java
@@ -1,13 +1,0 @@
-package interactions;
-
-import cards.Card;
-
-public abstract class View {
-    public abstract void print(String message);
-
-    public abstract void print(Card card);
-
-    public abstract void print(Card card1, Card card2);
-
-    public abstract void print(String[] list, String title);
-}
