--- player/AbstractPlayer.java
@@ -1,20 +1,0 @@
-package player;
-
-import java.util.List;
-
-import cards.Card;
-
-public abstract class AbstractPlayer {
-    protected Hand hand;
-    protected UsedPile usedPile;
-
-    public abstract void setCardToHand(List<Card> cardForPlayer);
-
-    public abstract void takeWonCard(Card card);
-
-    public abstract Card getTopCard();
-
-    public abstract int getUsedPileCount();
-
-    public abstract String getName();
-}
