--- player/Hand.java
@@ -1,30 +1,0 @@
-package player;
-
-import java.util.ArrayList;
-import java.util.List;
-import cards.Card;
-
-class Hand {
-    private List<Card> cardsToUse;
-    private UsedPile usedPileCards;
-
-    Hand(UsedPile usedPile){
-        cardsToUse = new ArrayList<Card>();
-        usedPileCards = usedPile;
-    }
-
-
-    void moveToUsedPile(Card card){
-        usedPileCards.addToUsedPile(card);
-    }
-
-    Card removeCard(){
-        Card returnCard = cardsToUse.get(0);
-        cardsToUse.remove(returnCard);
-        return returnCard;
-    }
-
-    void setListOfCards(List<Card> newListOfCard){
-        cardsToUse = newListOfCard;
-    }
-}
