--- player/UsedPile.java
@@ -1,22 +1,0 @@
-package player;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import cards.Card;
-
-class UsedPile{
-    private List<Card> usedCards;
-
-    UsedPile(){
-        usedCards = new ArrayList<Card>();
-    }
-
-    void addToUsedPile(Card newCard){
-        usedCards.add(newCard);
-    }
-
-    int getCount(){
-        return usedCards.size();
-    }
-}
