+++ src/main/resources/test.txt
@@ -1,0 +1,29 @@
+This is a test file
+However you might find something interesting here
+For example the sentence below seems far from nonsense
+This is nonsense
+Is it not
+what is love
+baby dont hurt me
+hurt me no more
+music music music
+oh no what is that
+
+There was no house now  no brewery  no building whatever left  but the
+wall of the old garden  The cleared space had been enclosed with a rough
+fence  and looking over it  I saw that some of the old ivy had struck
+root anew  and was growing green on low quiet mounds of ruin  A gate in
+the fence standing ajar  I pushed it open  and went in
+
+A cold silvery mist had veiled the afternoon  and the moon was not yet
+up to scatter it  But  the stars were shining beyond the mist  and the
+moon was coming  and the evening was not dark  I could trace out where
+every part of the old house had been  and where the brewery had been
+and where the gates  and where the casks  I had done so  and was looking
+along the desolate garden walk  when I beheld a solitary figure in it
+
+The figure showed itself aware of me  as I advanced  It had been moving
+towards me  but it stood still  As I drew nearer  I saw it to be the
+figure of a woman  As I drew nearer yet  it was about to turn away  when
+it stopped  and let me come up with it  Then  it faltered  as if much
+surprised  and uttered my name  and I cried out
