+++ src/main/resources/virus.csv
@@ -1,0 +1,12 @@
+name,type,stat1,stat2,stat3,stat4,stat5
+ebola,2,35600,15200,4,45,62
+corona,2,250500,7590,6,35,75
+influenza,3,50000000,5000000,3,31,9
+HIV,3,17000000,5800000,18,56,82
+smallpox,1,0,0,11,80,35
+rotavirus,3,1110000000,5920000,2,68,54
+herpes,3,46900500,480,4,49,37
+polio,1,9500,930,4,65,12
+syphilis,2,210000,15000,22,41,30
+marburg_virus,1,20,15,6,88,28
+human_rabies,2,590000,580000,30,89,82
